Split memory tests from Lib/test/test_capi/test_misc.py
